refactor(core): Remove legacy leader election and inline Lua-based implementation
Remove the old leader election strategy pattern and feature flag
(N8N_NEW_LEADER_ELECTION_IMPLEMENTATION). Inline the V2 Lua-script-based
implementation directly into MultiMainSetup, making it the only
implementation.
