+++ packages/@n8n/config/test/config.test.ts
@@ -374,7 +374,6 @@ describe('GlobalConfig', () => {
 			enabled: false,
 			ttl: 10,
 			interval: 3,
-			newLeaderElection: false,
 		},
 		generic: {
 			timezone: 'America/New_York',
