+++ packages/cli/src/scaling/multi-main-setup.ee.ts
@@ -1,18 +1,13 @@
-import { TypedEmitter } from '@/typed-emitter';
 import { Logger } from '@n8n/backend-common';
 import { GlobalConfig } from '@n8n/config';
 import { Time } from '@n8n/constants';
 import { MultiMainMetadata } from '@n8n/decorators';
 import { Container, Service } from '@n8n/di';
 import { ErrorReporter, InstanceSettings } from 'n8n-core';
+import assert from 'node:assert';
 
-import type * as LeaderElectionClientModule from '@/scaling/leader-election-client';
-import { Publisher } from '@/scaling/pubsub/publisher.service';
-import { RedisClientService } from '@/services/redis-client.service';
-
-import { MultiMainSetupLegacy } from './multi-main-setup-legacy';
-import type { MultiMainStrategy } from './multi-main-setup.types';
-import { MultiMainSetupV2 } from './multi-main-setup-v2';
+import { LeaderElectionClient } from '@/scaling/leader-election-client';
+import { TypedEmitter } from '@/typed-emitter';
 
 type MultiMainEvents = {
 	/**
@@ -33,9 +28,13 @@ type MultiMainEvents = {
 /** Designates leader and followers when running multiple main processes. */
 @Service()
 export class MultiMainSetup extends TypedEmitter<MultiMainEvents> {
-	private readonly strategy: MultiMainStrategy;
-
 	private leaderCheckInterval: NodeJS.Timeout | undefined;
+
+	private leaderCheckInProgress = false;
+
+	private get hostId() {
+		return this.instanceSettings.hostId;
+	}
 
 	constructor(
 		private readonly logger: Logger,
@@ -43,43 +42,26 @@ export class MultiMainSetup extends TypedEmitter<MultiMainEvents> {
 		private readonly globalConfig: GlobalConfig,
 		private readonly metadata: MultiMainMetadata,
 		private readonly errorReporter: ErrorReporter,
-		private readonly publisher: Publisher,
-		private readonly redisClientService: RedisClientService,
+		private readonly client: LeaderElectionClient,
 	) {
 		super();
 		this.logger = this.logger.scoped(['scaling', 'multi-main-setup']);
-
-		const emitFn = (event: 'leader-takeover' | 'leader-stepdown') => this.emit(event);
-
-		if (this.globalConfig.multiMainSetup.newLeaderElection) {
-			const { LeaderElectionClient } =
-				require('@/scaling/leader-election-client') as typeof LeaderElectionClientModule;
-			const client = Container.get(LeaderElectionClient);
-			this.strategy = new MultiMainSetupV2(
-				this.logger,
-				this.instanceSettings,
-				this.errorReporter,
-				client,
-				emitFn,
-			);
-		} else {
-			this.strategy = new MultiMainSetupLegacy(
-				this.logger,
-				this.instanceSettings,
-				this.publisher,
-				this.redisClientService,
-				this.globalConfig,
-				this.errorReporter,
-				emitFn,
-			);
-		}
 	}
 
 	async init() {
-		await this.strategy.init();
+		const result = await this.client.setLeaderIfNotExists();
+		if (!result.ok) {
+			this.logRedisCommandFailure('Failed to set leader key in Redis during init', result.error);
+			this.instanceSettings.markAsFollower();
+		} else if (result.result) {
+			// we became leader
+			this.takeOverAsLeader();
+		} else {
+			this.instanceSettings.markAsFollower();
+		}
 
 		this.leaderCheckInterval = setInterval(async () => {
-			await this.strategy.checkLeader();
+			await this.checkLeader();
 		}, this.globalConfig.multiMainSetup.interval * Time.seconds.toMilliseconds);
 	}
 
@@ -87,11 +69,21 @@ export class MultiMainSetup extends TypedEmitter<MultiMainEvents> {
 	async shutdown() {
 		clearInterval(this.leaderCheckInterval);
 
-		await this.strategy.shutdown();
+		if (this.instanceSettings.isLeader) {
+			// TODO: We should guard here that we only remove the key the key in Redis matches
+			// our host ID.
+			const result = await this.client.clearLeader();
+			if (!result.ok) {
+				this.logger.warn('Failed to clear leader key from Redis', { error: result.error });
+			}
+		}
+
+		this.client.destroy();
 	}
 
 	async fetchLeaderKey(): Promise<string | null> {
-		return await this.strategy.fetchLeaderKey();
+		const result = await this.client.getLeader();
+		return result.ok ? result.result : null;
 	}
 
 	registerEventHandlers() {
@@ -103,5 +95,130 @@ export class MultiMainSetup extends TypedEmitter<MultiMainEvents> {
 				return await instance[methodName].call(instance);
 			});
 		}
+	}
+
+	private async checkLeader() {
+		if (this.leaderCheckInProgress) {
+			this.logger.warn('Previous leader check is still in progress, skipping this check');
+			return;
+		}
+
+		this.leaderCheckInProgress = true;
+		try {
+			if (this.instanceSettings.isLeader) {
+				await this.checkAreWeStillLeader();
+			} else {
+				await this.checkCanBecomeLeader();
+			}
+		} finally {
+			this.leaderCheckInProgress = false;
+		}
+	}
+
+	/** Renew our leadership lease. If we've lost the lease, step down to follower. */
+	private async checkAreWeStillLeader() {
+		assert(this.instanceSettings.isLeader);
+
+		const renewTtlResult = await this.client.tryRenewLeaderTtl();
+		if (!renewTtlResult.ok) {
+			this.logRedisCommandFailure('Failed to renew leader TTL', renewTtlResult.error);
+			return;
+		}
+
+		const renewalResult = renewTtlResult.result;
+		if (renewalResult.id === 'success') {
+			this.logger.debug(`[Instance ID ${this.hostId}] Leader is this instance`);
+			return;
+		}
+
+		this.logger.warn('[Multi-main setup] Leader failed to renew leader key');
+
+		if (renewalResult.id === 'other-host-is-leader') {
+			this.logger.debug(
+				`[Instance ID ${this.hostId}] Leader is other instance "${renewalResult.currentLeaderId}"`,
+			);
+			this.stepDownToFollower();
+			return;
+		}
+
+		assert(renewalResult.id === 'key-missing');
+
+		const result = await this.client.setLeaderIfNotExists();
+		if (!result.ok) {
+			this.logRedisCommandFailure('Failed to set leader key in Redis', result.error);
+			this.stepDownToFollower();
+			return;
+		}
+
+		if (!result.result) {
+			this.stepDownToFollower();
+		}
+	}
+
+	private async checkCanBecomeLeader() {
+		assert(!this.instanceSettings.isLeader);
+
+		const getResult = await this.client.getLeader();
+		if (!getResult.ok) {
+			this.logRedisCommandFailure('Failed to get leader key from Redis', getResult.error);
+			return;
+		}
+
+		const leaderId = getResult.result;
+		if (leaderId && leaderId === this.hostId) {
+			this.errorReporter.info(
+				`[Instance ID ${this.hostId}] Remote/Local leadership mismatch, marking self as leader`,
+				{
+					shouldBeLogged: true,
+					shouldReport: true,
+				},
+			);
+
+			this.takeOverAsLeader();
+			return;
+		}
+
+		if (leaderId) {
+			this.logger.debug(`[Instance ID ${this.hostId}] Leader is other instance "${leaderId}"`);
+			return;
+		}
+
+		this.logger.debug(
+			`[Instance ID ${this.hostId}] Leadership vacant, attempting to become leader...`,
+		);
+
+		const result = await this.client.setLeaderIfNotExists();
+		if (!result.ok) {
+			this.logger.warn('Failed to try leader key set in Redis', { error: result.error });
+			return;
+		}
+
+		if (result.result) {
+			this.takeOverAsLeader();
+		}
+	}
+
+	private takeOverAsLeader() {
+		assert(!this.instanceSettings.isLeader);
+
+		this.logger.info(`[Instance ID ${this.hostId}] Leader is now this instance`);
+
+		this.instanceSettings.markAsLeader();
+
+		this.emit('leader-takeover');
+	}
+
+	private stepDownToFollower() {
+		assert(this.instanceSettings.isLeader);
+
+		this.logger.info(`[Instance ID ${this.hostId}] This is now a follower instance`);
+
+		this.instanceSettings.markAsFollower();
+
+		this.emit('leader-stepdown');
+	}
+
+	private logRedisCommandFailure(message: string, error: Error) {
+		this.logger.warn(`${message}: ${error.message}`, { error });
 	}
 }
