--- packages/cli/src/scaling/multi-main-setup.types.ts
@@ -1,6 +1,0 @@
-export interface MultiMainStrategy {
-	init(): Promise<void>;
-	shutdown(): Promise<void>;
-	checkLeader(): Promise<void>;
-	fetchLeaderKey(): Promise<string | null>;
-}
